library model corrected
